// src/modules/user/enums/user.messages.ts
export enum UserMessages {
  AlreadyExistsUser = 'User with this mobile already exists.',
  NotFoundUser = 'User not found',
  RemovedUserSuccess = 'User removed successfully',
  UpdatedUserSuccess = 'User Updated successfully.',
  MobileChangeLimit = 'You can only change your mobile number once every 24 hours.',
  VerifiedMobileSuccess = 'Mobile verified successfully.',
  AlreadyVerifiedMobile = 'Mobile already verified.',
  RevertedMobileSuccess = 'Reverted mobile successfully.',
  MobileVerifiedOrPrevNotFound = 'Mobile is verified or previous mobile not found.',
  CannotRemoveYourAccount = 'You cannot delete your account.',
  ChangedRoleSuccess = 'role changed successfully.',
  RoleAlreadyExistsInUser = 'This role already exists in user.',
  CannotChangeRole = 'You cannot change your role.',
  CannotChangeRoleToSuerAdmin = 'Cannot change role to super admin.',
}